Move logic outside of router
